+++ R/python_environment.R
@@ -708,7 +708,7 @@ checkPythonPackage <- function(package_name = NULL,
       errWidth = TRUE
     ))
   }
-  env_str_location <- stringr::str_locate(path_to_env, env_to_use)[2]
+  env_str_location <- GiottoUtils::string_locate(path_to_env, env_to_use)[2]
   # Change env_to_use from name of environment
   # to the full environment path
   env_to_use <- substr(path_to_env, 1, env_str_location)
@@ -745,8 +745,8 @@ checkPythonPackage <- function(package_name = NULL,
 
   contains_version_number <- grepl("==", package_name)
   if (contains_version_number) {
-    split_package_version <- stringr::str_split(package_name,
-      pattern = "=="
+    split_package_version <- strsplit(package_name,
+      split = "=="
     )
     package_name <- split_package_version[[1]][1]
     version_number <- split_package_version[[1]][2]